Now with proper math! Thanks to @Doc-Ok for the correction.
"If you want to map angles to linear distance for drawing concentric rings, you need to use the correct formula: x = z * tan(alpha), where z=10 units in your case. For 60 degrees that comes out to 17.320581 as expected, but for 30 degrees it's 5.7735 (instead of 8.6603...), etc. Your rings need to get wider the farther away from the center they are."